Solaris/Illumos build implementation proposal.

+++ report.c
@@ -136,7 +136,7 @@ void report_appendReport(pid_t pid, run_t* run, funcs_t* funcs, size_t funcCnt, 
     util_ssnprintf(run->report, sizeof(run->report), "DESCRIPTION: %s\n", description);
     util_ssnprintf(run->report, sizeof(run->report), "ORIG_FNAME: %s\n", run->dynfile->path);
     util_ssnprintf(run->report, sizeof(run->report), "FUZZ_FNAME: %s\n", run->crashFileName);
-    util_ssnprintf(run->report, sizeof(run->report), "PID: %d\n", pid);
+    util_ssnprintf(run->report, sizeof(run->report), "PID: %d\n", (int)pid);
     util_ssnprintf(
         run->report, sizeof(run->report), "SIGNAL: %s (%d)\n", util_sigName(signo), signo);
     util_ssnprintf(run->report, sizeof(run->report), "PC: 0x%" PRIx64 "\n", pc);
