+++ sanitizers.c
@@ -132,7 +132,7 @@ size_t sanitizers_parseReport(run_t* run, pid_t pid, funcs_t* funcs, uint64_t* p
     /* Under Linux the crash is seen in TID, but the sanitizer report is created for PID */
     pid = sanitizers_PidForTid(pid);
     snprintf(
-        crashReport, sizeof(crashReport), "%s/%s.%d", run->global->io.workDir, kLOGPREFIX, pid);
+        crashReport, sizeof(crashReport), "%s/%s.%d", run->global->io.workDir, kLOGPREFIX, (int)pid);
 
     FILE* fReport = fopen(crashReport, "rb");
     if (fReport == NULL) {
@@ -169,7 +169,7 @@ size_t sanitizers_parseReport(run_t* run, pid_t pid, funcs_t* funcs, uint64_t* p
             if (reportpid != pid) {
                 LOG_W(
                     "SAN report found in '%s', but its PID:%d is different from the needed PID:%d",
-                    crashReport, reportpid, pid);
+                    crashReport, reportpid, (int)pid);
                 break;
             }
             headerFound = true;
