+++ subproc.c
@@ -519,7 +519,7 @@ void subproc_checkTimeLimit(run_t* run) {
 
     if (run->tmOutSignaled && (diffUSecs > ((run->global->timing.tmOut + 1) * 1000000))) {
         /* Has this instance been already signaled due to timeout? Just, SIGKILL it */
-        LOG_W("pid=%d has already been signaled due to timeout. Killing it with SIGKILL", run->pid);
+        LOG_W("pid=%d has already been signaled due to timeout. Killing it with SIGKILL", (int)run->pid);
         kill(run->pid, SIGKILL);
         return;
     }
